+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - PostgreSQL: Display description of system variables
 - PostgreSQL: Avoid warning about crdb_version (bug #924, regression from 5.0.5)
 - PostgreSQL 11: Support PROCEDURE
+- SQLite PDO: Display results in SQL query
 - MS SQL: Fix collation issues when retrieving default values
 - CSS: Sticky table headers (bug #918)
 - CSS: Allow more custom styles with dark mode (bug #925)